Bug 1871545 - add a profiler counter for bandwidth use, r=profiler-reviewers,necko-reviewers,kershaw,canaltinova.
Differential Revision: https://phabricator.services.mozilla.com/D197148

+++ netwerk/base/nsUDPSocket.h
@@ -34,7 +34,7 @@ class nsUDPSocket final : public nsASocketHandler, public nsIUDPSocket {
   uint64_t ByteCountSent() override { return mByteWriteCount; }
   uint64_t ByteCountReceived() override { return mByteReadCount; }
 
-  void AddOutputBytes(uint64_t aBytes);
+  void AddOutputBytes(int32_t aBytes);
 
   nsUDPSocket();
